Extract constraint solver
